moved titlesortoptions into its own component file
