+++ blogproject/src/ViewDrafts.js
@@ -5,6 +5,7 @@ import AdvancedFormat from 'dayjs/plugin/advancedFormat'
 // import AdvancedFormat from 'dayjs/plugin/advancedFormat' // ES 2015
 import AlertText from './AlertText'
 import SearchOptions from './SearchOptions'
+import TitleSortOptions from './TitleSortOptions'
 
 
 
@@ -371,60 +372,7 @@ const DateSortOptions = (props)=>{
     )
 }
 
-const TitleSortOptions = (props)=>{
 
-    const handleTitleSort = (array, direction)=>{
-        let newArray
-        switch (direction){
-            case 'AZ' :
-            newArray = array.sort((a,b)=>{
-                if (a.title < b.title) {
-                    return -1;
-                  }
-                  if (a.title > b.title) {
-                    return 1;
-                  }
-                  return 0;
-            }) ;   
-            console.log(direction, newArray);
-            break;
-            case 'ZA' :
-                newArray = array.sort((a,b)=>{
-                    if (a.title > b.title) {
-                        return -1;
-                      }
-                      if (a.title < b.title) {
-                        return 1;
-                      }
-                      return 0;
-                }) ;   
-                console.log(direction, newArray);
-            break;
-            default:
-            console.log('no direction')
-        }
-        props.setNewArray(newArray)
-        props.toggleSort(false)
-    }
-
-    return(
-        <Container style={{border: '2px solid red', width: '50%'}}>
-        <p>By Title</p>
-        <Form.Group>
-        <Button
-        onClick={()=>{
-            handleTitleSort(props.array, 'AZ')
-        }}
-        >Title(AZ)</Button>
-        <Button
-        onClick={()=>{
-            handleTitleSort(props.array, 'ZA')
-        }}
-        >Title(ZA)</Button>
-        </Form.Group>
-        </Container>
-    )
-}
